Rewrite to allow reading the keydb without reading values into memory
The new implementation does not use msgpack anymore.

Using msgpack forced us to read the value while building the keydb at
startup.

While this could be workarounded by only encoding a header in msgpack
and then reading the Key,Value, it did not add much value.

See the README for the current limitations.